% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_excel_iknl.R
\name{arrange_iknl_tibble}
\alias{arrange_iknl_tibble}
\title{Arranges IKNL data in the order that is assumed for the calculation.R script to work}
\usage{
arrange_iknl_tibble(data, df_age_groups = AGE_GROUPS)
}
\arguments{
\item{data}{tibble - the output of \code{combine_excel_iknl}}

\item{df_age_groups}{tibble - columns \code{age_group} and \code{order_age_group} containing
the age group labels and order of age groups as they are assumed in the calculate.R script.
Refer to global_definitions.R}
}
\value{
tibble containing the same data as input data but then rearranged
}
\description{
Arranges IKNL data in the order that is assumed for the calculation.R script to work
}
