% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_excel_iknl.R
\name{combine_excel_iknl}
\alias{combine_excel_iknl}
\title{Combines excel files that are downloaded from IKNL}
\usage{
combine_excel_iknl(
  data_dir,
  outcome_var,
  cancer_sites,
  year_col,
  skip,
  sign = TRUE
)
}
\arguments{
\item{data_dir}{path where data is located}

\item{outcome_var}{string indicating the outcome variable "incidence" or "death"}

\item{cancer_sites}{tibble containing columns \code{cancer_site}, \code{abbreviation} and \code{order_site}. Refer to CANCER_SITES in global_definitions.R}

\item{year_col}{string indicating the column for the year under consideration}

\item{skip}{int - number of rows to skip}

\item{sign}{boolean - whether there is an asterisk after the year in the data, in which case it is removed from the data. Default is TRUE}
}
\value{
tibble containing the data in the excelsheets
}
\description{
see README for instructions on which data to download, how and where
}
