% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cancer_data.R
\name{convert_cancer_data}
\alias{convert_cancer_data}
\title{Convert cancer incidence and deaths data into arrays that are assumed in calculate.R}
\usage{
convert_cancer_data(
  .data,
  select_sex,
  select_cancer_site,
  value_col,
  df_age_groups = AGE_GROUPS
)
}
\arguments{
\item{.data}{tibble - output of \code{arrange_iknl_tibble}}

\item{select_sex}{character - indicating males or females \code{M} or \code{F}}

\item{select_cancer_site}{character - cancer site, refer to tibble in global_defintions.R}

\item{value_col}{character - column name of column containing the incidence or deaths numbers}

\item{df_age_groups}{tibble - columns \code{age_group} and \code{order_age_group} containing
the age group labels and order of age groups as they are assumed in the calculate.R script.
Refer to global_definitions.R}
}
\value{
array
}
\description{
Convert cancer incidence and deaths data into arrays that are assumed in calculate.R
}
