% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_age_groups.R
\name{make_age_groups}
\alias{make_age_groups}
\title{Helper function to make age groups from a column that contains age by year}
\usage{
make_age_groups(age_col_name, max_age = 80, age_group_band = 10)
}
\arguments{
\item{age_col_name}{data variable - name of the column that contains the
ages by year that should be assigned to age groups. Or vector with numeric
values that should be grouped into age groups}

\item{max_age}{int - lower bound of the oldest age group. Default 80 years,
i.e. the oldest age group is 80+}

\item{age_group_band}{int - width of the age group, default is 10 years, e.g.
age groups are 0-9, 10-19, etc}
}
\value{
vector with age groups
}
\description{
Helper function to make age groups from a column that contains age by year
}
